+++ src/config.ts
@@ -1,42 +1,31 @@
 import { workspace } from "vscode";
 
-type ConfigType = {
+export type ConfigType = {
   autoDetect: boolean;
   ignoreExtensions: string[];
   ignoreDir: string[];
   neededConvertCharset: string[];
-};
-
-export const config: ConfigType = {
-  // auto detect file encoding with GBK related.
-  autoDetect: true,
-
-  // ignore the specified file extensions, separated by comma.
-  ignoreExtensions: ["git", "ts", "vue"],
-
-  // ignore the specified directory, separated by comma.
-  ignoreDir: ["node_modules", ".vscode", ".idea", ".vscode-test", ".github"],
-
-  // Traditional and Simplified Chinese
-  neededConvertCharset: [
-    "Big5",
-    "GB2312",
-    "GB18030",
-    "EUC-TW",
-    "HZ-GB-2312",
-    "ISO-2022-CN",
-  ],
+  showBatchReport: boolean;
+  isBatch: boolean
 };
 
 // get user custom config
-export function getUserConfig() {
+export function getUserConfig(): ConfigType {
   const config = workspace.getConfiguration("GBK2UTF8");
   const _ignoreExt = config.get<string>("ignoreExtensions");
   const _ignoreDir = config.get<string>("ignoreDir");
 
   return {
+    // auto detect file encoding with GBK related.
     autoDetect: config.get<boolean>("autoDetect") as boolean,
-    ignoreExtensions: _ignoreExt ? _ignoreExt.split(",") : [],
+    // ignore the specified file extensions, separated by comma.
+    ignoreExtensions: _ignoreExt ? _ignoreExt.split(",") : config.ignoreExtensions,
+    // ignore the specified directory, separated by comma.
     ignoreDir: _ignoreDir ? _ignoreDir.split(",") : [],
+    // Traditional and Simplified Chinese
+    neededConvertCharset: ["Big5", "GB2312", "GB18030", "EUC-TW", "HZ-GB-2312", "ISO-2022-CN"],
+    // When batch convert encoding, show convert report result.
+    showBatchReport: config.get<boolean>("showBatchReport") as boolean,
+    isBatch: false,
   };
 }