+++ src/extension.ts
@@ -1,28 +1,18 @@
-import {
-  commands,
-  workspace,
-  window,
-  ExtensionContext,
-  Uri,
-  ProgressLocation,
-} from "vscode";
+import { commands, workspace, window, ExtensionContext, Uri, ProgressLocation } from "vscode";
 import * as fs from "fs";
 import * as path from "path";
 import * as iconv from "iconv-lite";
 import * as jschardet from "jschardet";
-import { config, getUserConfig } from "./config";
+import { ConfigType, getUserConfig } from "./config";
 
 const extensionName = "GBK2UTF8";
-let defaultConfig = config;
+let defaultConfig: ConfigType;
 
 export function deactivate() {}
 
 export function activate(context: ExtensionContext) {
-  defaultConfig = { ...defaultConfig, ...getUserConfig() };
-
-  context.subscriptions.push(
-    commands.registerCommand("GBK2UTF8.convert", convertWithProgress)
-  );
+  defaultConfig = getUserConfig();
+  context.subscriptions.push(commands.registerCommand("GBK2UTF8.convert", convertWithProgress));
 
   if (defaultConfig.autoDetect) {
     context.subscriptions.push(
@@ -47,14 +37,19 @@ async function convert(clickedFile: any, selectedFiles: any, progress?: any) {
     const filePaths = [];
 
     if (isDirectory) {
-      const files = fs.readdirSync(firstSelected.fsPath);
-      for (const fileName of files) {
-        const itemPath = path.resolve(firstSelected.fsPath, fileName);
-        const itemIsFile = fs.statSync(itemPath).isFile();
-        if (itemIsFile) {
-          filePaths.push(itemPath);
+      const iteratorFiles = (fsPath: string) => {
+        const _path = path.resolve(fsPath);
+        const isDir = fs.statSync(_path).isDirectory();
+        if (isDir) {
+          const dirs = fs.readdirSync(_path);
+          for (const fileName of dirs) {
+            iteratorFiles(path.resolve(_path, fileName));
+          }
+        } else {
+          filePaths.push(fsPath);
         }
-      }
+      };
+      iteratorFiles(firstSelected.fsPath);
     } else {
       for (const item of selectedFiles) {
         filePaths.push(item.fsPath);
@@ -65,6 +60,11 @@ async function convert(clickedFile: any, selectedFiles: any, progress?: any) {
       const item = filePaths[i];
       const uri = Uri.file(item);
       tasks.push(replaceContent(uri, true, progress));
+    }
+
+    // set config isBatch to true when have multiple files to convert
+    if (filePaths.length > 1) {
+      defaultConfig.isBatch = true;
     }
   }
 
@@ -83,7 +83,6 @@ function detectEncoding(fsPath: string) {
   fs.readSync(fd, sample, 0, sampleSize, null);
   fs.closeSync(fd);
   const detectedEncoding = jschardet.detect(sample);
-  console.log(detectedEncoding);
   return detectedEncoding;
 }
 
@@ -93,14 +92,9 @@ function detectEncoding(fsPath: string) {
  * @param encoding
  * @returns Promise<string>
  */
-function reEncodingContent(
-  filePath: string,
-  encoding: string
-): Promise<string> {
+function reEncodingContent(filePath: string, encoding: string): Promise<string> {
   return new Promise((resolve, reject) => {
-    const reader = fs
-      .createReadStream(filePath)
-      .pipe(iconv.decodeStream(encoding));
+    const reader = fs.createReadStream(filePath).pipe(iconv.decodeStream(encoding));
     const chunks: any[] = [];
     reader.on("error", () => {
       reject("oops, something went wrong!");
@@ -116,27 +110,32 @@ function reEncodingContent(
 }
 
 /**
- * replace the editor content with new encoding content
+ * replace content
  * @param uri
  * @param force
  * @param progress
  * @returns
  */
-async function replaceContent(
-  uri: Uri,
-  force: boolean = false,
-  progress?: any
-) {
+async function replaceContent(uri: Uri, force: boolean = false, progress?: any) {
+  defaultConfig = getUserConfig();
+
   const fsPath = uri.fsPath;
   let { encoding, confidence } = detectEncoding(fsPath);
   confidence = Number(confidence.toFixed(2));
 
-  if (config.neededConvertCharset.indexOf(encoding) === -1) {
-    if (force) {
+  const notChangedReturn = {
+    uri,
+    encoding,
+    confidence,
+    change: false,
+  };
+
+  if (defaultConfig.neededConvertCharset.indexOf(encoding) === -1) {
+    if (force && defaultConfig.isBatch) {
       const message = `It seems that the file encoding(${encoding}) is not GBK related.`;
       window.showWarningMessage(message);
     }
-    return { uri, encoding, confidence, change: false };
+    return notChangedReturn;
   }
 
   let dirIsIgnored = false;
@@ -147,16 +146,16 @@ async function replaceContent(
     }
   }
   if (dirIsIgnored && !force) {
-    return { uri, encoding, confidence, change: false };
+    return notChangedReturn;
   }
 
   const fileExt = fsPath.split(".").pop() || "";
   const fileName = uri.path.split("/").pop();
 
   if (defaultConfig.ignoreExtensions.includes(fileExt)) {
-    return { uri, encoding, confidence, change: false };
+    return notChangedReturn;
   }
-  const message = `It seems that the encoding of **${fileName}** is ${encoding}, do you want to convert it to UTF8?`;
+
   const writeToFile = async () => {
     const content = await reEncodingContent(fsPath, encoding);
     await workspace.fs.writeFile(uri, Buffer.from(content));
@@ -168,25 +167,28 @@ async function replaceContent(
     return await writeToFile();
   }
 
+  const message = `Seems the encoding of **${fileName}** is ${encoding}, do you want to convert it to UTF8?`;
   const answer = await window.showInformationMessage(message, "Yes", "No");
 
   if (answer === "Yes") {
     return await writeToFile();
   } else {
-    return { uri, encoding, confidence, change: false };
+    return notChangedReturn;
   }
 }
 
 function convertWithProgress(clickedFile: any, selectedFiles: any) {
   const options = {
-    location: ProgressLocation.Window,
-    cancellable: false,
+    location: ProgressLocation.Notification,
+    cancellable: true,
     title: "Convert",
   };
   window.withProgress(options, async (progress) => {
     progress.report({ message: "start convert..." });
     const result = await convert(clickedFile, selectedFiles, progress);
-    await writeLogFile(result);
+    if (defaultConfig.showBatchReport && defaultConfig.isBatch) {
+      await writeLogFile(result);
+    }
   });
 }
 
@@ -207,8 +209,7 @@ async function writeLogFile(result: LogItem[]) {
 
     writeContent += `\n## Converted (${convertedList.length})\n\n`;
 
-    const template = (item: LogItem) =>
-      `- \`[${item.encoding}/${item.confidence}]\` \`${item.uri.fsPath}\`\n`;
+    const template = (item: LogItem) => `- \`[${item.encoding}/${item.confidence}]\` \`${item.uri.fsPath}\`\n`;
 
     convertedList.forEach((item) => {
       writeContent += template(item);
@@ -221,9 +222,7 @@ async function writeLogFile(result: LogItem[]) {
     });
 
     const rootFolder = workspace.workspaceFolders[0];
-    const fileName = `${extensionName}-result-${Math.floor(
-      Math.random() * 100000
-    )}.md`;
+    const fileName = `${extensionName}-result-${Math.floor(Math.random() * 100000)}.md`;
     const filePath = rootFolder.uri.fsPath + `/${fileName}`;
     fs.writeFileSync(filePath, writeContent);